reorganize component file structure
