+++ client/src/components/CharacterNew/CharacterNew.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import Modal from "./UI/Modal/Modal"
-import Aux from "../hoc/Aux/Aux"
-import RaceInfo from "./RaceInfo/RaceInfo"
-import * as actions from '../actions'
+import Modal from "../UI/Modal/Modal";
+import Aux from "../../hoc/Aux/Aux";
+import RaceInfo from "../RaceInfo/RaceInfo";
+import * as actions from "../../actions";
 
 class CharacterNew extends Component {
   state = {
@@ -13,35 +13,33 @@ class CharacterNew extends Component {
   };
 
   removeModal = () => {
-    this.setState({showModal: false})
-  }
+    this.setState({ showModal: false });
+  };
 
-  selectRaceInfo = (race) => {
-    this.props.fetchRace(race)
+  selectRaceInfo = race => {
+    this.props.fetchRace(race);
     this.setState({
       showModal: true,
       selectedRace: race
-    })
-  }
+    });
+  };
 
   render() {
     return (
       <Aux>
-        <Modal
-          show={this.state.showModal}
-          removeModal={this.removeModal}
-        >
+        <Modal show={this.state.showModal} removeModal={this.removeModal}>
           <RaceInfo race={this.state.selectedRace} />
         </Modal>
 
         <div>
           <h1>CHOOSE YOUR RACE:</h1>
           <h6>Click a character race for more information</h6>
-          <ul>{this.state.races.map(race =>
-            <li key={race} onClick={() => this.selectRaceInfo(race)}>
-              {race}
-            </li>
-          )}
+          <ul>
+            {this.state.races.map(race => (
+              <li key={race} onClick={() => this.selectRaceInfo(race)}>
+                {race}
+              </li>
+            ))}
           </ul>
         </div>
       </Aux>
