+++ client/src/components/RaceInfo/RaceInfo.css
@@ -1,3 +1,7 @@
 .IMG {
   height: 200px;
 }
+
+.RaceInfoContainer {
+  width: 900px;
+}
