+++ client/src/components/UI/Modal/Modal.css
@@ -2,12 +2,12 @@
     position: fixed;
     z-index: 500;
     background-color: white;
-    width: 70%;
+    width: 80%;
     border: 1px solid #ccc;
     box-shadow: 1px 1px 1px black;
     padding: 16px;
     left: 15%;
-    top: 30%;
+    top: 15%;
     box-sizing: border-box;
     transition: all 0.4s ease-out;
 }
